Removendo arquivo .java código antigo e adicionado novas classes separadas
