+++ Alunos/Pedro Henrique/Aula-06/Cliente.java
@@ -1,0 +1,13 @@
+package grande_loja;
+
+public class Cliente {
+	String nome;
+    int idade;
+    String cidade;
+    String bairro;
+    String rua;
+
+    public void apresentarse() {
+        System.out.println("Nome: " + nome + ", Idade: " + idade);
+    }
+}
